Accordian typescript (#686)
* Accordian typescript

* Types are green!

* Accordian is no more

* removed unecessary type
