+++ libraries/core-react/src/Accordion/Accordion.types.ts
@@ -1,0 +1,6 @@
+export type AccordionProps = {
+  /** The header level, i.e. h1, h2, h3 etc. */
+  headerLevel?: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6'
+  /** Which side the chevron should be on  */
+  chevronPosition?: 'left' | 'right'
+}
