+++ libraries/core-react/src/Accordion/AccordionItem.tsx
@@ -1,13 +1,21 @@
-// @ts-nocheck
-import React, { forwardRef, useState } from 'react'
-import PropTypes from 'prop-types'
-import { commonPropTypes, commonDefaultProps } from './Accordion.propTypes'
+import React, { forwardRef, useState, ReactElement } from 'react'
+import type { AccordionProps } from './Accordion.types'
 
-const AccordionItem = forwardRef(function AccordionItem(
+type Props = {
+  index?: number
+  accordionId?: string
+  /** Is AccordionItem expanded */
+  isExpanded?: boolean
+  /** accordion item is disabled */
+  disabled?: boolean
+} & JSX.IntrinsicElements['div'] &
+  AccordionProps
+
+const AccordionItem = forwardRef<HTMLDivElement, Props>(function AccordionItem(
   {
     headerLevel,
     chevronPosition,
-    index,
+    index = 0,
     accordionId,
     isExpanded,
     children,
@@ -27,7 +35,7 @@ const AccordionItem = forwardRef(function AccordionItem(
     const panelId = `${accordionId}-panel-${index + 1}`
 
     return childIndex === 0
-      ? React.cloneElement(child, {
+      ? React.cloneElement(child as ReactElement, {
           isExpanded: expanded,
           toggleExpanded,
           id: headerId,
@@ -37,7 +45,7 @@ const AccordionItem = forwardRef(function AccordionItem(
           parentIndex: index,
           disabled,
         })
-      : React.cloneElement(child, {
+      : React.cloneElement(child as ReactElement, {
           hidden: !expanded,
           id: panelId,
           headerId,
@@ -52,27 +60,5 @@ const AccordionItem = forwardRef(function AccordionItem(
 })
 
 AccordionItem.displayName = 'eds-accordion-item'
-
-AccordionItem.propTypes = {
-  ...commonPropTypes,
-  /** @ignore */
-  accordionId: PropTypes.string,
-  /** Is AccordionItem expanded */
-  isExpanded: PropTypes.bool,
-  /** @ignore */
-  index: PropTypes.number,
-  /** @ignore */
-  children: PropTypes.node.isRequired,
-  /** accordion item is disabled */
-  disabled: PropTypes.bool,
-}
-
-AccordionItem.defaultProps = {
-  ...commonDefaultProps,
-  disabled: false,
-  index: 0,
-  accordionId: '',
-  isExpanded: false,
-}
 
 export { AccordionItem }
