+++ libraries/core-react/src/Accordion/index.ts
@@ -1,8 +1,16 @@
-// @ts-nocheck
-import { Accordion } from './Accordion'
+import { Accordion as BaseAccordion } from './Accordion'
 import { AccordionItem } from './AccordionItem'
 import { AccordionHeaderTitle, AccordionHeader } from './AccordionHeader'
 import { AccordionPanel } from './AccordionPanel'
+
+type AccordionProps = typeof BaseAccordion & {
+  AccordionItem: typeof AccordionItem
+  AccordionHeader: typeof AccordionHeader
+  AccordionHeaderTitle: typeof AccordionHeaderTitle
+  AccordionPanel: typeof AccordionPanel
+}
+
+const Accordion = BaseAccordion as AccordionProps
 
 Accordion.AccordionItem = AccordionItem
 Accordion.AccordionHeader = AccordionHeader
